:recycle: Tests of Instrumentation.Instrument and CanInstrument combined, separated by type of member being instrumented
